Blog Based on Jekyll
Added About and changed a bit format of posts, as I will be the only one posting.

Things todo will be on the todo list always.

I don't like that it pulls the entirety of fontawesome to work, but I guess is needed for a static website to work properly.